feat: add exclusion fn
